fix tests and demos

+++ rubato/struct/gameobject/physics/hitbox.py
@@ -714,13 +714,13 @@ class Circle(Hitbox):
         if self.color:
             relative_pos = camera.transform(self.pos)
             scaled_rad = camera.scale(self.radius)
-            Draw.queue_circle(relative_pos, scaled_rad, self.color, fill=self.color, z_index=self.true_z)
+            Draw.queue_circle(relative_pos, int(scaled_rad), self.color, fill=self.color, z_index=self.true_z)
 
         if self.debug or Game.debug:
             if not relative_pos:
                 relative_pos = camera.transform(self.pos)
                 scaled_rad = camera.scale(self.radius)
-            Draw.queue_circle(relative_pos, scaled_rad, Color(0, 255), int(2 * Display.display_ratio.x))
+            Draw.queue_circle(relative_pos, int(scaled_rad), Color(0, 255), int(2 * Display.display_ratio.x))
 
     def clone(self) -> Circle:
         return Circle(
